Moved inotify into osutil
