+++ osutil/inotify/PATENTS
